Updated the db writer so that the column names are added to the file

+++ node_src/ethoscope_node/utils/mysql_db_writer.py
@@ -37,18 +37,20 @@ class MySQLdbCSVWriter(object):
             self._dst_path=dst_path
 
             self._csv_file_name = self._dst_path+"/"+self._remote_db_name + ".txt"
-            #print ("Filename:", self._csv_file_name)
+            print ("Filename:", self._csv_file_name)
 
             #TODO add dir exists check and create if needed
-
-            if overwrite:
-                logging.info("Trying to remove old database")
-                try:
-                    os.remove(self._csv_file_name)
-                    logging.info("Success")
-                except OSError as e:
-                    logging.warning(e)
-                    pass
+            #Only try to remove file if it already exists
+            if os.path.isfile(self._csv_file_name):
+                print ("Filename:", self._csv_file_name)
+                if overwrite:
+                    logging.info("Trying to remove old database")
+                    try:
+                        os.remove(self._csv_file_name)
+                        logging.info("Success")
+                    except OSError as e:
+                        logging.warning(e)
+                        pass
 
 
 
@@ -57,11 +59,8 @@ class MySQLdbCSVWriter(object):
 
     def update_roi_tables(self):
         """
-        Fetch new ROI tables and new data points in the remote and use them to update local db
-
-        :return:
+        Fetch new ROI tables from mysql database and populate in a textfile
         """
-
         src = MySQLdb.connect(host=self._remote_host, user=self._remote_user,
                                          passwd=self._remote_pass, db=self._remote_db_name)
 
@@ -79,16 +78,30 @@ class MySQLdbCSVWriter(object):
         src.close()
 
     def _update_one_roi_table(self, table_name, roi_num,src):
+        """
+        Fetch the data from a specified table and write to file.
+        """
         src_cur = src.cursor()
-        src_command = "SELECT * FROM %s" % (table_name)
-        src_cur.execute(src_command)
-        #TODO add column names to first row
-        for sc in src_cur:
-            with open(self._csv_file_name,"a") as f:
-                row = "\t".join(["{0}".format(val) for val in sc])
-                row += "\t"+str(roi_num)
-                f.write(row)
-                f.write("\n")
+        #Grab the col names
+
+        try:
+            src_command = "SELECT * FROM %s" % (table_name)
+            src_cur.execute(src_command)
+            num_fields = len(src_cur.description)
+            field_names = [i[0] for i in src_cur.description]
+            #print "Field Names: ", field_names
+            row_names = "\t".join([name for name in field_names])
+            row_names += "\t"+"roi"+"\n"
+            for sc in src_cur:
+                with open(self._csv_file_name,"a") as f:
+                    row = "\t".join(["{0}".format(val) for val in sc])
+                    row += "\t"+str(roi_num)
+                    f.write(row)
+                    f.write("\n")
+        except TypeError as e:
+            logging.warning(e)
+            pass
+
 
     def enumerate_roi_tables(self):
         """
@@ -105,11 +118,17 @@ class MySQLdbCSVWriter(object):
         cur = src.cursor()
         cur.execute(command)
         rois_in_src = set([c[0] for c in cur])
+        rowiter = 0
         for i in rois_in_src :
             src_cur = src.cursor()
             src_command = "SELECT * FROM ROI_%i" % (i)
             src_cur.execute(src_command)
-            #TODO add column names to first row
+            if rowiter == 0:
+                field_names = [i[0] for i in src_cur.description]
+                row_names = "\t".join([name for name in field_names])
+                row_names += "\t"+"roi"+"\n"
+                rowiter = rowiter + 1
+                yield row_names
             for sc in src_cur:
                 row = "\t".join(["{0}".format(val) for val in sc])
                 row += "\t"+str(i)+"\n"
